+++ tests/Mock/InterfaceDependencyRequiredArguments.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Micro
+ *
+ * @copyright   Copryright (c) 2015-2018 gyselroth GmbH (https://gyselroth.com)
+ * @license     MIT https://opensource.org/licenses/MIT
+ */
+
+namespace Micro\Container\Testsuite\Mock;
+
+class InterfaceDependencyRequiredArguments
+{
+    protected $foo;
+
+    public function __construct(StringArgumentsInterface $foo)
+    {
+        $this->foo = $foo;
+    }
+
+    public function getDependency(): StringArgumentsInterface
+    {
+        return $this->foo;
+    }
+
+    public function getFoo(): string
+    {
+        return $this->foo->getFoo();
+    }
+}