+++ tests/Mock/Simple.php
@@ -8,4 +8,13 @@ declare(strict_types=1);
  * @copyright   Copryright (c) 2015-2018 gyselroth GmbH (https://gyselroth.com)
  * @license     MIT https://opensource.org/licenses/MIT
  */
-require_once './vendor/autoload.php';
+
+namespace Micro\Container\Testsuite\Mock;
+
+class Simple
+{
+    public function test(): bool
+    {
+        return true;
+    }
+}