+++ tests/Mock/StringArgumentsChild.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Micro
+ *
+ * @copyright   Copryright (c) 2015-2018 gyselroth GmbH (https://gyselroth.com)
+ * @license     MIT https://opensource.org/licenses/MIT
+ */
+
+namespace Micro\Container\Testsuite\Mock;
+
+class StringArgumentsChild extends StringArguments implements StringArgumentsInterface
+{
+    public function isChild(): bool
+    {
+        return true;
+    }
+}