+++ tests/Mock/StringArgumentsInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Micro
+ *
+ * @copyright   Copryright (c) 2015-2018 gyselroth GmbH (https://gyselroth.com)
+ * @license     MIT https://opensource.org/licenses/MIT
+ */
+
+namespace Micro\Container\Testsuite\Mock;
+
+interface StringArgumentsInterface
+{
+    public function isChild(): bool;
+}